requirements.txt
Django==5.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
python-dotenv==1.0.0
mysql-connector-python==8.0.33